+++ core/src/main/java/org/mule/runtime/core/internal/management/stats/visitor/InputDecoratorVisitor.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.internal.management.stats;
+package org.mule.runtime.core.internal.management.stats.visitor;
 
 import java.io.InputStream;
 import java.util.Collection;
@@ -14,14 +14,14 @@ import java.util.Set;
 
 import org.mule.runtime.api.streaming.bytes.CursorStreamProvider;
 import org.mule.runtime.core.api.management.stats.CursorComponentDecoratorFactory;
-import org.mule.runtime.core.internal.management.stats.InputDecoratorVisitor.Builder;
+import org.mule.runtime.core.internal.management.stats.InputDecoratedCursorStreamProvider;
 
 /**
- * Visitor that returns a decorator for output statistics.
+ * Visitor that returns a decorator for input statistics.
  * 
  * @since 4.4, 4.3.1
  */
-public class OutputDecoratorVisitor implements Visitor {
+public class InputDecoratorVisitor<T> implements Visitor<T> {
 
   private final CursorComponentDecoratorFactory decoratorFactory;
   private final String correlationId;
@@ -30,9 +30,12 @@ public class OutputDecoratorVisitor implements Visitor {
   private boolean decorateCollections = true;
   private boolean decorateCursorProviders = true;
 
-  private OutputDecoratorVisitor(CursorComponentDecoratorFactory decoratorFactory, String correlationId,
-                                 boolean decorateInputStreams, boolean decorateIterators, boolean decorateCollections,
-                                 boolean decorateCursorProviders) {
+  private InputDecoratorVisitor(CursorComponentDecoratorFactory decoratorFactory,
+                                String correlationId,
+                                boolean decorateInputStreams,
+                                boolean decorateIterators,
+                                boolean decorateCollections,
+                                boolean decorateCursorProviders) {
     this.decoratorFactory = decoratorFactory;
     this.correlationId = correlationId;
     this.decorateInputStreams = decorateInputStreams;
@@ -43,32 +46,55 @@ public class OutputDecoratorVisitor implements Visitor {
 
   @Override
   public InputStream visitInputStream(VisitableInputStream visitable) {
-    return decoratorFactory.decorateOutput(visitable.getDelegate(), correlationId);
+    if (decorateInputStreams) {
+      return decoratorFactory.decorateInput(visitable, correlationId);
+    } else {
+      return visitable.getDelegate();
+    }
   }
 
   @Override
-  public Iterator visitIterator(VisitableIterator visitable) {
-    return decoratorFactory.decorateOutput(visitable.getDelegate(), correlationId);
+  public Iterator<T> visitIterator(VisitableIterator visitable) {
+    if (decorateIterators) {
+      return decoratorFactory.decorateInput(visitable.getDelegate(), correlationId);
+    } else {
+      return visitable.getDelegate();
+    }
   }
 
   @Override
-  public Collection visitCollection(VisitableCollection visitable) {
-    return decoratorFactory.decorateOutputResultCollection(visitable.getDelegate(), correlationId);
+  public Collection<T> visitCollection(VisitableCollection visitable) {
+    if (decorateCollections) {
+      return decoratorFactory.decorateInput(visitable.getDelegate(), correlationId);
+    } else {
+      return visitable.getDelegate();
+    }
   }
 
   @Override
-  public List visitList(VisitableList visitableList) {
-    return visitableList;
+  public List<T> visitList(VisitableList visitableList) {
+    if (decorateCollections) {
+      return (List) decoratorFactory.decorateInput(visitableList, correlationId);
+    }
+    return visitableList.getDelegate();
   }
 
   @Override
-  public Set visitSet(VisitableSet visitableSet) {
-    return visitableSet;
+  public Set<T> visitSet(VisitableSet visitableSet) {
+    if (decorateCollections) {
+      return (Set) decoratorFactory.decorateInput(visitableSet, correlationId);
+    }
+    return visitableSet.getDelegate();
   }
 
+
   @Override
-  public CursorStreamProvider visitCursorStreamProvider(VisitableCursorStreamProvider visitableCursorStreamProvider) {
-    return visitableCursorStreamProvider;
+  public CursorStreamProvider visitCursorStreamProvider(VisitableCursorStreamProvider cursorStreamProvider) {
+    if (decorateCursorProviders) {
+      return new InputDecoratedCursorStreamProvider(cursorStreamProvider.getDelegate(), decoratorFactory, correlationId);
+    }
+
+    return cursorStreamProvider.getDelegate();
   }
 
   public static Builder builder() {
@@ -117,9 +143,9 @@ public class OutputDecoratorVisitor implements Visitor {
       return this;
     }
 
-    public OutputDecoratorVisitor build() {
-      return new OutputDecoratorVisitor(factory, correlationId, decorateInputStreams, decorateIterators, decorateCollections,
-                                        decorateCursorProviders);
+    public InputDecoratorVisitor build() {
+      return new InputDecoratorVisitor(factory, correlationId, decorateInputStreams, decorateIterators, decorateCollections,
+                                       decorateCursorProviders);
     }
 
   }
