+++ core/src/main/java/org/mule/runtime/core/internal/management/stats/visitor/VisitableInputStream.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.internal.management.stats;
+package org.mule.runtime.core.internal.management.stats.visitor;
 
 import java.io.IOException;
 import java.io.InputStream;
